moved tests to dir
