+++ test/unit/specs/components/ImageDetails/image-info.spec.js
@@ -1,5 +1,5 @@
 import ImageInfo from '@/components/ImageDetails/ImageInfo';
-import render from '../../test-utils/render';
+import render from '../../../test-utils/render';
 
 describe('Image Info', () => {
   let props = null;
