+++ test/unit/specs/components/ImageDetails/reuse-survey.spec.js
@@ -1,6 +1,6 @@
 import ReuseSurvey from '@/components/ImageDetails/ReuseSurvey';
 import { DETAIL_PAGE_EVENTS, SEND_DETAIL_PAGE_EVENT } from '@/store/usage-data-analytics-types';
-import render from '../../test-utils/render';
+import render from '../../../test-utils/render';
 
 describe('ImageAttribution', () => {
   let options = null;
