+++ test/unit/specs/components/ImageDetails/social-share-buttons.spec.js
@@ -1,7 +1,7 @@
 import SocialShareButtons from '@/components/ImageDetails/SocialShareButtons';
 import { SOCIAL_MEDIA_SHARE } from '@/store/action-types';
 import { SEND_DETAIL_PAGE_EVENT, DETAIL_PAGE_EVENTS } from '@/store/usage-data-analytics-types';
-import render from '../../test-utils/render';
+import render from '../../../test-utils/render';
 
 describe('SocialShareButtons', () => {
   let options = null;
